[refactor] タイルを抽象クラスベースに
