+++ Assets/Scripts/Enviroment.cs
@@ -1,0 +1,7 @@
+/**
+ * 環境情報を保持するクラス
+ */
+public static class Environment
+{
+    public const int TileSize = 1;
+}
