+++ Assets/Scripts/Enviroment.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c3d6d2e772b34c999c11a6840f7a62da
+timeCreated: 1720360520
