+++ Assets/Scripts/InvasionPhase/InvasionController.cs
@@ -8,10 +8,14 @@ namespace InvasionPhase
         /** シーン間のデータ共有オブジェクト */
         [SerializeField] private CreateToInvasionData createToInvasionData;
 
+        /** 迷路作成等を行うコントローラ */
+        [SerializeField] private InvasionMazeController mazeController;
+
         // Start is called before the first frame update
         public void Start()
         {
             Debug.Log(createToInvasionData);
+            mazeController.Create(createToInvasionData.TileData, createToInvasionData.TrapData);
         }
     }
 }
